testing/requirements.txt
# Testing Framework Dependencies
# NUAH Trading Agent Test Suite

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# HTTP client
requests>=2.31.0
httpx>=0.25.0

# Async support
asyncio-throttle>=1.0.2
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
faker>=22.0.0
tqdm>=4.66.0

# Visualization & Reporting
matplotlib>=3.8.0
tabulate>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Date/time handling
python-dateutil>=2.8.2

